(TK-364) Another jruby-pool-int-test race condition cleanup attempt
This commit makes another attempt at cleaning up some transient race
condition failures which appear to be due to some tests misinterpreting
that a pool flush has completed when it is still in progress and,
therefore, failing when attempting to assert that fresh instances have
been put back into the pool.

Previous code added a watch on the pool agent and interpreted the
receipt of a callback on the watch as an indication that a flush had
succeeded.  It seemed possible - although not locally reproducible -
that the callback could be invoked prematurely.  This commit changes the
code to perform an await on the pool agent after the point that the
flush is in progress, assuming that the await will block until after the
flush should have completed.