(TK-367) Add JRubyInstanceSchema schema and integration test
The initialization functions users of this library add will most likely
associate things onto the JRubyInstance record. While this fulfills the
`instance?` relationship (i.e. the record with added keys is still an
instance of the JRubyInstance record), it will no longer satisfy the
JRubyInstance schema.

This commit adds an additional JRubyInstanceSchema schema that has all the
keys that must be a part of the record, as well as a `schema/Keyword
schema/Any` so that users can add whatever additional keys they would like.

Also add an integration test that tests adding custom initialization and
cleanup functions.