(TK-367) Remove mocking of instance creation
Previously, `create-pool-instance!` was mocked out using `with-redefs` in many
tests. This was done primarily in order to speed up the tests, because
previously the real `create-pool-instance!` loaded all the puppet code.
However, that no longer happens by default - only if an initialization
function is defined to do that.

This commit removes `create-mock-pool-instance!` and accompanying
`with-redefs` fixture. Doing so increases the time of `lein test` by ~3
seconds when run on my laptop, but that seems like a worthwhile tradeoff for
much easier to understand and refactor tests.