(TK-367) Pass full jruby config to initialize-scripting-container
Previously, `:initialize-scripting-container` was named
`:initialize-env-variables` and was only used as a hook for setting
environment variables, and thus was only passed `gem-home` as a variable to
use. However, since it has been renamed and can be used for any initialization
of the scripting container before we call `.runScriptlet` on it, it makes more
sense to pass it the full JRuby config.