(TK-367) Remove JRubyInstanceSchema
Previously, we had a `JRubyInstanceSchema` that had all the keys of the
JRubyInstance record plus `schema/Keyword schema/Any` so that we could assoc
additional keys onto the record. However, this was non-ideal because it might
be possible to update the record to require additional keys but not update
this schema. It turns out that `schema/defrecord` has an `extra-key-schema?`
to which we can supply `{schema/Keyword schema/Any}`, causing it to do exactly
what we want! With this, assoc-ing an additional key to the record does not
produce a schema error, but removing a required key does.