(TK-366) Remove JRuby TK service
Remove the jruby-pool TK service and protocol. Instead of this service, users
of this library will use the `pool-manager-service` to create a pool context,
and then pass that pool context to the functions in the `jruby-core`
namespace, to borrow, return, and flush instances.

Move the `with-lock` and `with-jruby-instance` macros out of the service and
into the `jruby-core` namespace. Add a core `register-event-handler` function.

Add a test TK service to the `jruby-service-test` namespace to test that a TK
service that depends on the new pool-manager-service and uses its
`create-pool` works as expected. Replace the
`test-pool-population-during-init` test in this namespace with unit tests -
the logic of this test is specific to puppet server, and there should be an
integration test in the puppet server repo for puppet server's logic. The unit
tests added test that a borrow will block on the pool being initialized and
that a borrow will block until an instance becomes available.