(TK-370) Add `flush-instance!` service function
Add a `flush-instance!` function to the pool manager service, to allow for
flushing of a single jruby instance.

In addition, this function is now passed into `prime-pool!` in the
`create-pool` service function, so that it is registered on each instance as
the `flush-instance-fn`, used whenever the instance is flushed.

Also, fix a small alignment error in th earguments to the defservice.