(TK-370) Remove flush functions from pool manager service
Previously, the pool manager service provided functions for creating a pool,
flushing an instance from the pool, flushing the entire pool, and flushing the
pool for shutdown (flushing and not creating new instances).

This commit removes all of these flush functions, leaving the pool manager
service with a single `create-pool` function. This was done because the state
that is needed for flushing is contained in the pool context and in the
instances themselves. Having the state be also in the service - or in the
service sometimes, and in the pool context/instances other times - is like
having one foot in each world. Furthermore, it was agreed that it is
preferable to have an API that has all of the state contained in the pool
context and pass that around into plain, non-TK functions than to have to use
TK functions for everything or to have to pass the TK service/context around
everywhere. Given that we will still control the creation of the pool, we can
add what we need to the pool to coordinate flushing across multiple pools.

This commit also adds wrappers around the `jruby-agents/send-flush-*!`
functions in the `core` namespace, so that callers don't need to know anything
about the agents. However, `flush-instance!` function was not added to the
core namespace, as we do not yet know exactly what the use case will be for
that - we might end up separating out the `max-requests` logic, and then a
`flush-instance` function would become more useful, but we aren't sure about
that yet.