(TK-366) Remove top-level :jruby key
Previously, `create-pool` required as input a config without a top level
`:jruby` key, but `initialize-config` took in a map with a top level `:jruby`
key and returned a map without this key. This commit updates
`initialize-config` to need a top level `:jruby` key, so that it is consistent
with `create-pool`.

In addition, update `cli-run!` to not call `initialize-config`, and require
the config passed into `cli-run!` and `cli-ruby!` to match the JRubyConfig
schema. This matches the behavior of `create-pool` - thus any time we
instantiate JRuby containers, we require users to pass in a pre-munged config
that matches the schema. Users can use `initialize-config` to munge the config
and provide defaults, or they can do their own munging.