(main) Add a foo.rb file and have cli tests use it to do a require
Previously there were several tests commented out because the required puppet
to be on the ruby load path, and we don't want puppet to be a dependency (even
a test one) of this library. This commit adds a `foo.rb` ruby file with a
single method, `foo` that returns the string `bar`. This file can then be
called from the tests to make sure that the `ruby` and `irb` commands can
require a library on the load path.