(TK-366) Add event-callbacks to pool-context
Add the event-callbacks atom to the pool-context so that the
`with-jruby-instance` and `with-lock` macros do not require it to be passed
in.

Note that the other core functions - e.g. `borrow-from-pool(-with-timeout)`,
`return-to-pool`, 'lock', 'unlock', and the flush functions were not updated
to not have `event-callbacks` as an arg because most of these do not take
`pool-context` as an arg, and they already required `event-callbacks` as an
arg, whereas the macros previously existed in the service and pulled the
`event-callbacks` from the service context.