(TK-376) Add :internal key to pool context and instance
Both the pool context and the instance contain keys that we don't want users
to mess with - such as the pool and flush agents and the `flush-instance-fn`.
This commit moves all keys we want to keep "private" in both the pool context
and the instance underneath an `:internal` key.

For the the pool context, the `:config` key remains public; for the instance,
the `:id` and `:scripting-container` remain public.