(maint) Add timed-await to jruby-agents-test
The jruby-agents-test namespace was failing in travis with

```
No output has been received in the last 10 minutes, this potentially indicates
a stalled build or something wrong with the build itself.
```

which means that possibly an `await` never actually worked.

This commit moves the `timed-await` helper to `jruby-testutils` and updates
jruby-agents-test to use it instead of `await`.

Also, remove the unused `wait-for-jrubies` test helper.