(TK-379) Move non-public functions out of jruby-core
Previously, the `jruby-core` namespace held a combination of public and
non-public functions. We want to isolate the public functions of this library
to one namespace + the service functions. This commit moves all non-public
functions out of `jruby-core` (and moves some functions in other namespaces
that we probably want to be public to `jruby-core`) to achieve this.