(SERVER-1408) Update to PL release of JRuby 1.7.26
This commit updates us to use a PL release of JRuby 1.7.26, as
there are upstream fixes that we really need and it's not clear
when they might be doing their own release.

The commit also updates a few tests, because recent changes in
JRuby introduce a 'RUBY' environment variable into all
ScriptingContainers.