(TK-398) Fix with-jruby-retry-test race condition
This commit adds a borrow to the main service pool in the
`with-jruby-retry-test-via-mock-get-pool` function before any borrows
with intentional retries are attempted.  Previously, without the extra
borrow, the test's use of the `with-redefs` function to switch out the
`get-pool` function could cause pool instances to only be created in
the retry pool and not the main service pool, leading a borrow attempt
made against the main service pool to hang.