(SERVER-797) Extract an interface for ScriptingContainer
This commit extracts an interface for use with schema/type validation
with ScriptingContainers.  This helps establish the baseline methods
that external consumers of the library can expect to be present for any
ScriptingContainers that the library creates for them.  The interface
is named ScriptingContainer.  The corresponding implementation class is
renamed to InternalScriptingContainer.