(PE-13552) Refactor jruby pool to use one pool
This commit modifies the flush logic to avoid the need for two pools,
which would get swapped out when a flush occured.

It also removes the pool-agent and relies on the locking the pool
to resolve multiple flush requests happening at once.

The flush-instance-agent now takes the form of a modify-instance-agent,
so that any operations which create or destroy a jruby must go through
this agent