(SERVER-1475) Add an JRuby slf4j Logger adapter
This commit adds a class, Slf4jLogger, which implements the JRuby Logger
interface, forwarding any logging calls which are received on to an
slf4j logger.  This class doesn't actually hook up the logger to created
JRuby instances since JRuby classes typically get the object that they
use for logging during static initialization, which would most likely
complete before any attempt to set a custom logger with JRuby could be
done.  Use of this Logger would typically require that the consuming
process set the jruby.logger.class Java system property to the namespace
of the new class on the command line, e.g.,
-Djruby.logger.class=com.puppetlabs.jruby_utils.jruby.Slf4jLogger.