(PE-13552) Fill pool with poison pills
Instead of adding only one pill. This will allow the collect-all-jrubies
fn to lock the pool in the event that 2 shutdown requests happen one
after the other. Previously, the second shutdown flush would becomes deadlocked
trying to get the pool lock, as it would wait for the pool to become filled
before getting the lock