(SERVER-1630) Add logic to cleanup jar files on container termination
This commit adds some hacky logic to the InternalScriptingContainer
class to cleanup jar files and associated file descriptors when the
container is terminated.  This code is compensating for work that will
hopefully be done in JRuby itself at some point, as part of a resolution
for https://github.com/jruby/jruby/issues/3928.