(SERVER-1630) Add deprecation warning for compat-version with JRuby 9k
This commit adds a deprecation warning for the case that the
compat-version setting requested in the JRuby config cannot be used
because the JRuby libraries being used are supporting a different
language version.  This warning would be displayed if a compat-version
value of 1.9 or 2.0 were requested but JRuby 9k, which has a fixed
version newer than 2.0, were in use.