(SERVER-1630) Explicitly terminate jruby containers on test exit
This commit reworks all tests which create jruby ScriptingContainers to
ensure that the `terminate` method is explicitly called for each
container as the test finishes.  Without this cleanup, tests could
periodically fail with jar loading exceptions, most likely due to low
memory conditions.  Explicit termination ensures that all memory
associated with a container is made eligible for garbage collection as
soon as the associated test finishes, as opposed to when the process may
be running low on available heap.