(SERVER-1630) Changed compat-version warnings and logging
Previously, the `compat-version` logged as part of the jruby config for
JRuby 9k listed a misleading value of either 1.9 or 2.0.  In this
commit, the `compat-version` should always list the actual language
version used, for either JRuby 1.7 or 9k.

This commit adds a helper function, using-jruby-9k?, for determining
whether the current JRuby version in use is a 9k-based version (as
opposed to a JRuby 1.x version, for now).

This commit also adds some unit tests related to the `compat-version`
handling, including tests which validate that the proper warnings are
generated when an unsupported version is specified and/or the version is
not configurable.