(SERVER-1661) Remove support for compat-version setting
This commit removes support for the compat-version setting from
jruby-utils.  The compat-version setting was introduced in jruby-utils
version 0.4.0 to allow for JRuby 1.7 to optionally run under Ruby
language version 2.0 instead of the default version, 1.9.  JRuby
1.7.27 effectively breaks the ability to use Ruby language version
2.0, however, due to a regression - see:
https://github.com/jruby/jruby/issues/4613 - and is unlikely to be
fixed in the future.  JRuby 9k also does not support compat-version
configurability.  Now that jruby-utils can be used with
JRuby 9k for Ruby language 2+ support, it is no longer valuable to be
able to set compat-version to 2 for use with JRuby 1.7.

This commit effectively fixes JRuby 1.7 to always use Ruby language
version 1.9 - as was the case before jruby-utils version 0.4.0.
jruby-utils will now have to be used with JRuby 9k for Ruby language
2+ version support.