(SERVER-1780) Enable jruby profiling for ScriptingContainers
In order to allow users to use the built in jruby profiler and collect
the profiling output, this commit adds two new configuration options to
the jruby containers. profiling-mode can be used to select the type of
the jruby profiler to use, and it defaults to off. profiler-output-file
can be used to set the desired output file to be used to log profiler
output. If profiler-output-file is not set it will default to a randomly
named file relative to the current working directory of the service.