(SERVER-2060) Splay JRuby instance flushing
Previously all JRuby instances would be flushed after they reached
`max-requests-per-instance`. This could cause a thundering herd of
JRuby instance initialization which we've observed during
performance testing.

This patch creates a `splay-instance-flush` configuration option
(available via code for testing but not intended to be plumbed through
to the user). If this is true (the default value) when we first
initialize a JRuby Pool and fill it with instances each instance
receives an `initial-borrows` value that is a fraction of the
max-requests. If set, the `return-to-pool` method will prefer the
value of initial-borrows to that of max-requests to determine if
the instance should be flushed. Subsequent instance creations wil not
receive an initial-borrows value unless the entire pool is being
flushed and refilled.