(SERVER-2081) Set JRuby KCode, source encoding, and external encoding
This commit adds a function that can configure the KCode, external
encoding, and source encoding of the JRuby scripting containers. We use
this function when initializing our JRuby instances to set these
encodings to UTF-8, to avoid parsing source files as US-ASCII, which can
cause issues with translation interpolation.

Note that this is only relevant in JRuby 1.7; in JRuby 9K, source files are parsed
as UTF-8 by default.

This is a cherry-pick from the master branch.