(SERVER-1777) Do not override JRuby ScriptingContainter.terminate
Previously, we manually introspected JRuby's classloading and tempdir
handling to ensure that libraries copied to a tempdir when loaded would
be cleaned up on JRuby instance termination (upstream registers them as
deleteOnExit of the JVM).

With
https://github.com/jruby/jruby/commit/6493b3a93dcca2f9e8fac746cee8cf807594d161
JRuby now removes these temporary jars when the per-instance classloader
is terminated. This fix is available in JRuby 9.1.14.0 and later and
supersedes our code.