(maint) Remove redundant formatting from logging
This commit swaps formatting log methods such as `log/infof` for their
unformatted counterparts: `log/info`. This is done in cases where the
log method is receiving a value generated by `i18n`. The `i18n` methods
perform formatting, which renders the additional formatting done by
`log` redundant and potentially dangerous as any translation that
generates a sequence of characters that looks like a format statement
will cause an exception to be raised.