(SERVER-2529) Create ReferencePool
This commit adds a new implementation of the LockablePool interface,
that only maintains one JRuby instance and simply hands out multiple
references to it on a borrow request. The pool can be locked and
unlocked just like the JRubyPool.

The main semantic difference is that the pool can only be cleared
(and the instance unregistered) when all borrows have been returned.
This prevents us from deleting a JRuby instance out from under a
thread that is still using it. The changes necessary to accomodate this
in the Clojure code are in the commits that follow.