(SERVER-2529) Clojure changes for toggling JRuby pool implementation
This commit contains all of the necessary clojure changes to toggle and
interact with the new ReferencePool. Notable changes are:

1) Which pool is used can be toggled via the `jruby-puppet.multithreaded`
setting in the tk config.

2) The addition of a JRubyPool protocol that defines different methods
for clearing an individual instance and for clearing the whole pool.
This was necessary because the ReferencePool cannot meaningfully clear a
single instance -- clearing the instance is more or less equivalent to
clearing the whole pool, and therefore blocks on all the borrowed
references being returned. The new protocol handles this polymorphism.