(SERVER-2529) Refactor flush to allow extra borrow count check
This commit refactors the `flush-pool` protocol function for the
ReferencePool to isolate the actual flushing logic from the locking
logic. This allows to check the borrow count both before and after
locking in the case that we are flushing due to max-borrows being hit,
to avoid duplicate flushes. The `flush-pool` protocol method locks
around the flushing logic without checking any additional conditions,
and is used when flushing the pool via the admin API.