(SERVER-2529) Put watcher on borrow-count atom rather than agent
Previously, the flushing tests were watching for changes in the
modify-instance-agent, which syncrhonizes changes to the pool. But a
recent refactor caused actions to be sent to the agent whether or not a
flush actually occurred. So this commit updates the watcher to instead
watch the `borrow-count` atom and only deliver a promise if the
borrow-count went down, i.e. got reset during a flush.