(SERVER-2529) Make shutdown lock the RefPool
Previously, we weren't locking the ReferencePool while shutting it down.
This meant that it was possible for the pool to be flushed during
shutdown, which caused a race condition where we might try to clean up a
null instance. This commit updates the shutdown function to lock the
pool prior to cleaning up its instance. It then inserts the shutdown
pill before unlocking, to ensure any pending borrows or locks will fail
with the appropriate error message.