(SERVER-2685) Respect `flush-timeout` setting in ReferencePool
This commit updates all functions that involve flushing the JRuby ReferencePool
to time out if the set `flush-timeout` time has elapsed. In most cases, we throw
a TimeoutException, except in the `return` function where we just log that the
timeout occurred. This is due to `flush-if-at-max-borrows` being called by an
agent, which doesn't propagate exceptions very well. This maintains the same
behavior as the existing InstancePool.