(SERVER-2681) Add `availableJrubies` method
This commit adds a method for getting the number of free Jrubies in the pool.
This is the same as `currentSize` for the existing instance pool, but for the
reference pool (multithreaded) this will either be 1 (if nothing is using the
pool) or 0. This ensures that the `num-free-jrubies` metric (which uses
`free-instance-count`) is accurate for both pool types.