(SERVER-2715) Record worker-id when borrowing pool resources
This commit adds a `:worker-id` event field to be sent to the metrics service.
For the instance pool, this is the instance id we've already been using. For
the new reference pool, this is the thread id. This will allow us to reason
about pool usage in similar ways for both pool types.