(SERVER-2679) Ensure return callback runs before actual return
This commit puts the `instance-returned` callback before the actual `return`
again in the public-facting `return-to-pool` method. Previously, if a request
was waiting for a new instance, calling `return` first could trigger a new
borrow so quickly that the borrow event callback would trigger before the return
event callback. For the instance pool, because instances ids are reused, instead
of 1. removing the instance from the `borrowed-instances` metric and then 2.
adding it back with the new borrow event, we ended up 1. updating the
`borrowed-instances` with the new event and then 2. removing the instance from
`borrowed-instances`, which caused the metric to lose track of a whole instance.