(SERVER-3098) Log pool lock events at INFO level
The JRuby pool lock is used to gain exclusive access to the JRuby
pool. When requested, threads will be prevented from borrowing
a JRuby instance until all instances are returned to the pool and
the lock is released.

This behavior gives the JRuby lock lifecyle of request, acquire, release
an extremely critical impact on JRuby throughput. This change increses
the log statements around these lifecycle events to INFO level so
that post-mortem debugging is possible when services using the JRuby
pool are not logging at DEBUG level.