(PE-36375) Take up JRuby 9.4.3
When we upgraded to JRuby 9.4 we broke our IRB testing because of an
issue in JRuby and the io/console library. The FFI that io/console
integrates with has changed and the 'path' variable no longer exists on
linux based consoles unless the IO is backed by an actual file.[1]

This fails in our testing where we stub stdin & stdout with essentially
a StringIO object, but does not fail when running IRB normally with an
actual terminal attached.

This should have a known issue written about it, however I know of no
use case where a user would write a Puppet extension that needed an
interactive console.

This bug was also released in Puppet Server 8.0 and thus far we have not
heard of issues regarding it so we have anecdotal data that Puppet users
are not needing programmatic access to an interactive console via JRuby.

The fix for this issue has been committed[2] to the io/console repository
but has not been released yet. Once io/console >= 0.6.0 has been
released and JRuby has taken it up, our test changes and any known
issues around this should be resolved.

1. https://github.com/jruby/jruby/issues/7727
2. https://github.com/ruby/io-console/commit/f1bb90fe01b6d9b5db68437ffd4acf7c47695168