(PE-37376) Initialize one JRuby instance first when initializing pool
Puppet may apply a "settings catalog" when it initializes. This will
check the filesystem and create any directories specified by settings
that do not already exist. This will cause errors if Puppet does need to
create directories and is initialized in parallel. To resolve this when
priming the pool we now instantiate one instance first before
initializing the remaining instances in parallel.